feat(protocol): use sequencing system
IDs now must arrive in the correct sequence. A new EXPECT flag can be
set on a packet to inform the sender that the receiver was expecting a
different sequence number to the one it received. IDs are also now
queued using a channel so that the send method must wait until an ID is
available before sending.